Add metrics update_op to Estimator's eval_metric_ops.
Before the new API for exporting SavedModels in different modes came along, the only way get evaluation metrics was to call Estimator.evaluate, which would run the metric ops via SessionRunHooks. Now, the evaluation metrics can be exported as part of the SavedModel, when saved in EVAL mode, so we need to include the update_op in the eval metrics.

PiperOrigin-RevId: 203133707