Raise UserWarning when trying to call model_fn directly in train mode.
Training the adanet.Estimator requires calling its `train` method. Under the hood, `train` calls `model_fn` in a while loop in order to incrementally grow the graph, one iteration at a time.

The internal model_fn, which is accessible via the `model_fn` property on the Estimator, should not be called in TRAIN mode outside the context of the while loop. Since there is no mechanism for for the graph to grow, the AdaNet behavior will no longer be correct.

However, calling `model_fn` directly in EVAL or PREDICT mode should be okay since they do not require modifying the graph.

PiperOrigin-RevId: 207295505