Pass iteration step `Tensor` to base learners.
This `Tensor` provides a more accurate idea of the relative step since the beginning of the iteration than the global step, and eliminates the need for users to create their own step-tracking variable.

For example, `iteration_step` can be used for learning rate decay schedules when passed as the `global_step` argument.

PiperOrigin-RevId: 207333374