Zero-debias adanet loss moving average.
Previously, the adanet_loss used for tracking the current best candidate was biased towards zero in the beginning due to the way the moving average was computed: the first value in the moving average was zero, and subsequent adanet_loss values were averaged with it.

This change corrects the adanet_loss moving average to be more correct in TensorBoard, but should have no effect on the training.

Rename `adanet_loss_ema` to simple `adanet_loss` in TensorBoard.

Also identify off-by-one training error in TODO.

PiperOrigin-RevId: 208085897