Fix off-by-one error in iteration training steps.
Previously, an iteration would end one step too early. The fix was to pass the iteration step which is zero-indexed, instead of one-indexed, to the candidate. This had the additional benefit of removing a redundant `tf.Variable` that was duplicating the work of tracking the current iteration step.

Also eliminate non-determinism in reading the value of iteration step and global step. These are now incremented after all the train ops are complete.

PiperOrigin-RevId: 208482865