Replace tf.stack(metric_ops) with tf.group(metric_ops).
A user might supply a metric tuple where the second element is a valid
update op, but cannot be converted to a Tensor. This previously caused
AdaNet to crash, and this change fixes this.

Also modifies ReportMaterializer to run the metric update ops before
evaluating the metric Tensors, rather than evaluating the metric update
ops and using the values as the metric in a single step.

PiperOrigin-RevId: 208829093