Report ensemble architecture as an eval metric.
This makes it easier to see at a glance in both the commandline and TensorBoard which ensembles are currently being considered, and which are doing the best at each evaluation.

This change requires storing the name of `BaseLearner` as a `Tensor` property on `WeightedBaseLearner`.

Also update report_accessor to properly handle Python3 serialized proto values (as `bytes`).

PiperOrigin-RevId: 213653300