Rebuild the ensemble across iterations without a frozen graph.
In order to realign AdaNet’s SoT to the TensorFlow SoT, AdaNet must not rely on an exported GraphDef for constructing the graph across iterations. This design introduces a way to efficiently grow the graph by calling only the python code that originally defined it and restoring learned variables from the checkpoint. This change will be completely invisible to the end-user.

# Detailed Design

The key observation is that AdaNet has access to all the Python code it needs to create the graph at each iteration, no matter what iteration it is in.

Currently during the prepare_next_iteration phase at the end of iteration t, the AdaNet framework will look up the frozen graph for iteration t-1 on disk. This graph contains the ops of the previous_ensemble including the best candidates and their mixture weights.

However, this is unnecessary because the framework has access to the user’s subnetwork.Generator and subnetwork.Builders at each iteration. It can call the subnetwork.Generator t times in order to get the subnetwork.Builders, and reconstruct the ensemble every time. Doing so eliminates the freezer (and associated phase) altogether, because the graph pruning can occur during the checkpoint-overwrite phase directly, hence “freezerless”. This is a huge benefit, since the freezer code is brittle and has been the source of many bugs. In fact, most of the core TensorFlow library code it depends on for pruning, such as tf.graph_util.convert_variables_to_constants, isn’t very well maintained to handle its many edge cases.

Pruning the graph of extraneous candidates that were not chosen to be part of the final ensemble requires storing the names of the subnetworks that were the best at each iteration. That way AdaNet only needs to call the Builder.build_subnetwork method to construct the graph of the subnetworks that are in the previous_ensemble.

Once all the required builders are selected, the next thing to do is restore the learned weights of the previous ensemble. It turns out that this is very simple: after creating the previous ensemble (blue in the above diagram), we can call tf.train.warm_start() to restore all variables from the checkpoint before overwriting it. Apart from being a core TensorFlow tool, it will check that all variables that it attempts to restore are located in the checkpoint, as an added sanity check.

## Benefits

 * Eliminates Freezer: The freezer code is the most brittle code inside AdaNet, and it is slow. There are many edge cases it needs to handle such as variable batch size across modes, different features per mode, different devices per run, collocation constraints.

 * The freezer phase is slow due to I/O: it needs to read and write a large proto (the GraphDef) to disk which is even slower when it needs to be written to CNS.

 * Ability to share information across iterations without Tensors: Since everything now happens in a single graph by calling Python APIs, we can extend Subnetworks to allow users to share arbitrary data across iterations including Tensors, Python primitives and objects, and even lambdas. This opens up many possibilities that were not previously possible in the Freezer-based approach.

 * Eases Report materialization: Users who don’t provide a report materializer can now have access to the hparams used in previous runs without needing to write anything to disk. This speeds up training, and make extending the ReportMaterializer more straightforward.

 * Better cross device compatibility: By creating the graph dynamically at each iteration, AdaNet is more robust to changes in devices. Training can begin on GPU, and then continue on CPU or TPU because the only thing that is shared between runs is the checkpointed variables. Likewise, models trained on one device, can be rebuilt entirely on a different device for serving, allowing for better graph optimizations than if the frozen graph were imported.

 * More flexible device placement capability: Creating the graph means more flexibility with allocating previously trained subgraphs to different devices. This can be used to provide model-parallelism to the ensemble while different subnetworks train on their respective devices (for a future design).

## Caveats

One assumption this requires is that calls to the subnetwork.Generator’s must be idempotent: given the same arguments, it must produce the same candidates which produce the same subgraphs. The good news is that this restriction was already present in order to enable distributed training: If candidate and their graphs were not generated deterministically, workers would attempt to restore variables that were absent from the checkpoint and crash. Note that initializers might be different since they are only called once when there is no checkpoint to restore from. This caveat should be kept in mind as we implement smart generators that risk producing different results whenever generate_candidates is called.

PiperOrigin-RevId: 223008101