Introduce `adanet.ensemble.SoloStrategy` to produce single subnetwork models.
This effectively allows AdaNet to do architecture search without needing to produce an ensemble model. Occasionally, a single subnetwork may perform better than an ensemble, in particular when other subnetworks in the ensemble perform poorly enough to harm overall performance. This strategy can be used in conjunction with other ensemble strategies in order to find the best possible model structure, in case the best model at a given iteration is just a single subnetwork.

PiperOrigin-RevId: 236174051