Create new experimental `adanet.distributed` package for distributed training.
Introduce `drop_remainder` option to `adanet.distributed.RoundRobinStrategy` in order to only have one subnetwork per subnetwork-worker, regardless of the number of candidate subnetworks and workers. This can be used during training when each subnetwork assumes it is the only one in the graph, and the presence of other subnetworks would cause errors.

No longer execute training hooks for subnetworks when not training them.

Spawn evaluator task in distributed training tests.

PiperOrigin-RevId: 236701739