Support bootstrap aggregating (a.k.a. bagging) in `AutoEnsembleEstimator`.
Introduces the `AutoEnsembleSubestimator` object which specifies which training data to train its corresponding `tf.estimator.Estimator` instance on.

Internally, it calls the `tf.estimator.Estimator`'s model_fn within a `tf.template` to create the model_fn's trainable variables once and share them between calls. This model_fn is called once on the feature and labels from the input_fn passed to `Estimator#train()` in order to obtain logits for the Ensembler, and once again for the `train_input_fn` when present in order to obtain the correct train op.

PiperOrigin-RevId: 248439862