Dedicate a step variable per subnetwork and ensemble.
This is primarily useful for the round-robin placement strategy to allow subnetworks with different training rates to complete at their own pace. Furthermore, it eliminates contention on incrementing the global step in distributed training: each subnetwork only  increments its own step variable, and the chief is solely responsible for assigning the global step.

This will also enable subnetworks to train for different numbers of steps within the same iteration, and monitor the steps per subnetwork in TensorBoard.

Finally, the global step is now a function of the current subnetwork iteration steps. By default, it is the max of all subnetwork steps, but can be any monotonically increasing function such as sum depending on the application.

PiperOrigin-RevId: 257444217