Execute train ops in hooks for CPU and GPU.
This will allow AdaNet to stop training subnetworks and ensembles after a specified numbers of steps or seconds, or when encountering NaN loss during training, without needed to terminate the current TensorFlow session and reconstruct the graph.

Furthermore, this change simplifies subnetwork and ensemble training management by allowing it to all happen in python code instead of TensorFlow ops. No longer do we need a tf.Variable to monitor the training state of the iteration; instead we use a _TrainManager object that tracks the training state of each candidate. Each candidate has a dedicated SessionRunHook that updates the _TrainManager state when it is done training.

On TPU, uneven training is also supported, but in a less granular manner since `SessionRunHooks` are only run every `num_iterations_per_loop`. Note that train ops and steps-incrementation happen together as a single op. Also instead of using a `tf.Variable` to track when the iteration is over, it uses the _TrainManager object.

PiperOrigin-RevId: 260213039