Gracefully handle NaN losses from candidate ensembles and subnetworks during training.
This can occur when a candidate is generated with a pathological hyperparameter set, such as using a learning rate that is too high and causes training instability.

Internally on CPU and GPU, when an ensemble or subnetwork has a NaN loss during training, its training is marked as terminated, and its train op is no longer run on subsequent calls to `session.run`. On TPU, the NaN loss candidates' train ops will still execute, but should have no effect. In all cases, as long as one ensemble plus its underlying subnetworks do not have a NaN loss, training will continue.

PiperOrigin-RevId: 260218536