Remove TF_CONFIG patching since this causes issues with distributed training.
Up until now, AdaNet has patched the TF_CONFIG to remove the parameter servers
and workers during the graph growing phase, since this phase only happens locally
on the chief. However, this seems to cause some issues with the way stateful optimizers
create the scope of partitioned variables.

This change removes the patching and always keeps the cluster around even during
the growing phase. This means that during the growing phase we will incur some
RPC overheads from talking to the cluster, however the bottleneck in AdaNet is not
the growing phase anyways. These overheads should be ok.

PiperOrigin-RevId: 261115556