Delegate calls to train, evaluate, predict, and export to a temporary Estimator.
Previously, AdaNet Estimators delegated calls to their super. But any time we wanted to pass additional arguments to the adanet model_fn during training, we needed to store them as fields on the instance. Tracking what value a field has at any given time in the training lifecycle was not clear.

This change gives us the flexibility to pass arbitrary params to the model_fn of the new Estimator, and reduce state stored inside the AdaNet Estimator.

PiperOrigin-RevId: 262022201