Run the `Evaluator` before `evaluate`, `predict`, and `export_saved_model`.
**BREAKING CHANGE**: This CL may affect models that use an adanet.Evaluator, by choosing different candidates at the last AdaNet iteration. This changes more closely matches the expected behavior when using passing an Evaluator to the adanet.Estimator or adanet.AutoEnsembleEstimator constructor.

When the user passes an Evaluator, we run it to establish the best candidate during evaluation, predict, and export_saved_model. Previously they used the adanet_loss moving average collected during training. While the previous ensemble would have been established by the Evaluator, the current set of candidate ensembles that were not done training would be considered according to the adanet_loss.

Now when a user passes an Evaluator that, for example, uses a hold-out set, AdaNet runs it before making predictions or exporting a SavedModel to use the best new candidate according to the hold-out set.

PiperOrigin-RevId: 262223244