Disable asserts in v1 binary classifier _Head during estimator.evaluate().
AdaNet and AutoEnsembleEstimator sometimes have a few NaN and non-NaN subnetworks at a given iteration. This doesn't crash during training, since AdaNet simply chooses the best subnetwork among the non-NaN candidates. However, during estimator.evaluate(), AdaNet evaluates all subnetworks and ensembles. This triggers an assert failure in V1 binary classifier _Head, since it expects the predictions to be between 0 and 1, and NaN is not between 0 and 1. This causes AdaNet and AutoEnsembleEstimator to raise an exception during estimator.evaluate(), even though the actual final model is servable. Hence, we disable this bounds check.

PiperOrigin-RevId: 270316326