Make model_fn construction explicit.
This makes the code more functional and breaks our dependence on having to rely
on the params dictionary for arbitrary arguments and makes it explicit what the
supported parameters are.

This is also safer as the code will break if we try to pass in unsupported
arguments or some arguments are misspelled (params dict would not catch these).

PiperOrigin-RevId: 282053712