Preserved the outputs in the PredictionOutput that are not in the best_export_outputs.
The current implementation will overwrite the outputs for ALL signatures that have a PredictionOutput to the best_export_outputs. In fact, some user specified signatures are not intended to be used for label prediction, whose values should be preserved.

A fix is to compare the keys in the old outputs and the best_export_outputs. If they happen to be the same, then update them. To avoid side-effect and maximum compatibility, this CL is to simply combine the original outputs with the best_export_outputs and use the latter if keys are the same.

PiperOrigin-RevId: 292968149